requirements.txt
schedule
beautifulsoup4
pymongo
python-dateutil
